Starting to refactor all the contain logic into a separate object, it is big enough to be its own class and was strating to get hard to follow
